+++ CHANGES.rst
@@ -123,6 +123,8 @@ Bug Fixes
 - ``astropy.modeling``
 
   - Added missing default values for ``Ellipse2D`` parameters. [#3903]
+  - Improved support for pickling of compound models, including both compound
+    model instances, and new compound model classes. [#3867]
 
 - ``astropy.nddata``